+++ nlppln/wfgenerator.py
@@ -9,3 +9,15 @@ class WorkflowGenerator(WFGenerator):
         steps_dir = os.path.join(module_path, 'cwl')
 
         WFGenerator.__init__(self, steps_dir=steps_dir)
+
+    def save(self, fname, inline=True, relative=False, validate=True,
+             encoding='utf-8'):
+        """Save workflow to file
+
+        For nlppln, the default is to save steps inline.
+        """
+        super(WorkflowGenerator, self).save(fname,
+                                            inline=inline,
+                                            relative=relative,
+                                            validate=validate,
+                                            encoding=encoding)